+++ app/src/main/java/com/example/batteryservicekotlin/database/BatteryDao.kt
@@ -15,7 +15,7 @@ interface BatteryDao {
     // Запрос выборки записанной в течение выбранных суток
     //startDay - начало дня, endDay - конец дня
     @Query("SELECT * FROM unit WHERE date BETWEEN :startDay AND :endDay")
-    fun getChosenDayUnits(startDay: Int, endDay: Int): LiveData<List<Unit>>
+    fun getChosenDayUnits(startDay: Long, endDay: Long): LiveData<List<Unit>>
 
     @Query("SELECT date FROM unit")
     fun getDates(): LiveData<List<Date>>
